requirements.txt
gymnasium===0.26.2
pygame
numpy===1.26.2
opencv-python
matplotlib
pillow
tensorflow
keras
tqdm

